+++ frontend/src/pages/Home/PostListPage/PostListPage.stories.tsx
