+++ frontend/src/pages/Home/index.tsx
@@ -1,6 +1,7 @@
 import AppInstallPrompt from '@components/common/AppInstallPrompt';
 import Layout from '@components/common/Layout';
-import PostListPage from '@components/post/PostListPage';
+
+import PostListPage from './PostListPage';
 
 export default function Home() {
   return (
